requirements.txt
langchain-mistralai
langchain-xai
yt-dlp
requests
langgraph
fastapi==0.115.12
uvicorn==0.34.2
pydantic
SQLAlchemy==2.0.40
psycopg2-binary
alembic==1.15.2
python-dotenv
uuid
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
authlib
httpx
feedparser
beautifulsoup4
celery
flower
apscheduler